Expose SFGatewayMock as part of the library to facilitate testing
